Move intended public api to top level namespace
No need for it to live under Server, it doens't really describe
anything about the related methods and nows our chance to sneak in
some breaking changes before we actually release the gem.

I might use the now free Server class to house some of the actual
server code so it can be split from the remote drb object.