Support configurable allowed requests
We have a scenario where we want particular domains to be permitted
through the proxy as we want a real response rather than a mock. Since
MiniProxy defaults to responding with an empty 200 when an unmocked
request is made, we can't do that.

This adds an 'allow_request' option that takes a method and URL pattern.
When a request is made that matches an allowed request it will be
permitted through the proxy.

Requests for localhost, 127.0.0.1, and whatever the MiniProxy configured
host is will continue to always be permitted.