Fix proxy method extensions in Ruby 2.6
In Ruby 2.6 the signature of `perform_proxy_request` has changed; it no
longer takes two arguments and a block but now three or four arguments.
This means the PUT, PATCH, and DELETE extensions made to the WEBrick
HTTP proxy server do not work on Ruby 2.6.

This adds test coverage for each supported HTTP method and fixes the
extensions.

https://github.com/ruby/webrick/blob/4938ec3fae3ce32cceb685f6b8efff8926340c09/lib/webrick/httpproxy.rb#L302